requirements.txt
html5lib==1.1
importlib_metadata==7.1.0
isodate==0.6.1
owlrl==6.0.2
packaging==24.0
prettytable==3.10.0
pyparsing==3.1.2
pyshacl==0.25.0
PyYAML==6.0.1
rdflib==7.0.0
six==1.16.0
wcwidth==0.2.13
webencodings==0.5.1
zipp==3.18.1
